division into classes for windows
