+++ Program/source/my_profile.json
@@ -1,0 +1,1 @@
+{"name": null, "gender": 0, "height": null, "weight": null, "birthday": null, "record_100": null, "record_200": null, "record_400": null, "record_800": null, "record_1_000": null, "record_3_000": null, "record_5_000": null, "record_10_000": null, "record_21_000": null, "record_42_000": null}
